Version 11.8.22.1 (cherry-pick)
Merged 3409e1d19ed484cef48e09f48f6d6f7b3a25a3c1

[sandbox] testing.cc: Add explicit conversion for SIGSTKSZ

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4775020
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/11.8.22@{#2}
Cr-Branched-From: a01e3a51b69354635767ce59a7410d585a904e4f-refs/heads/main@{#89496}

+++ src/sandbox/testing.cc
@@ -493,7 +493,7 @@ void SandboxTesting::InstallSandboxCrashFilter() {
   stack_t signalstack = {
       .ss_sp = alternate_stack,
       .ss_flags = 0,
-      .ss_size = SIGSTKSZ,
+      .ss_size = static_cast<size_t>(SIGSTKSZ),
   };
   CHECK_EQ(sigaltstack(&signalstack, nullptr), 0);
